wip: rename InputFieldBorder -> BorderEffect + cleanup  + add comments / use new textfield styling
